Fix training for roberta

+++ model/masked_bert.py
@@ -24,6 +24,6 @@ class MLMBert(TrainConfig):
             input_ids=input_ids,
             attention_mask=attention_mask,
             token_type_ids=token_type_ids,
-            labels=labels.type(torch.LongTensor),
+            labels=labels,
         )
